requirements.txt
Django==5.2.3
djangorestframework==3.16.0
validators==0.35.0
requests==2.32.4
beautifulsoup4==4.13.4
tldextract==5.3.0
psycopg2-binary==2.9.10
pytest-django==4.11.1
pytest==8.4.1
